Added a placeholder classifier model
Trained a quick and dirty placeholder model to use in development.  Won't be accurate at all but it should have the correct format for output.  Too large to commit to github.com but added it to the gitignore

Will still need to figure out how to handle cropping/resizing images though.